fastfetch

+++ local_to_workflow.sh
@@ -1,19 +1,22 @@
 clear
-echo "Local files going to workflow! Ctrl+C to cancel!"
-echo -n "3..."
-sleep 1
-echo -n "2..."
-sleep 1
-echo -n "1..."
-sleep 1
+read -p "Local to Workflow? [y/N] " confirm
+confirm=${confirm,,}  # Convert to lowercase
+
+if [[ "$confirm" != "y" && "$confirm" != "yes" ]]; then
+    echo "Maybe not."
+    exit 1
+fi
 rm -r  fonts/
 cp -r ~/.local/share/fonts/ fonts/
 rm -r  .config/
-cp -r ~/.config/starship/ .config/
-cp -r ~/.config/rofi/ .config/
+mkdir .config/
 cp -r ~/.config/waybar/ .config/
 cp -r ~/.config/kitty/ .config/
 cp -r ~/.config/fish/ .config/
 cp -r ~/.config/niri/ .config/
+cp -r ~/.config/rofi/ .config/
+mkdir .config/starship/
+cp -r ~/.config/starship/starship.toml .config/starship/
+mkdir .config/fastfetch/
+cp ~/.config/fastfetch/config.jsonc .config/fastfetch/
 echo "Done!"
-sleep 2
